Split integration test root
