+++ tests/integration/discovery.rs
@@ -1,0 +1,23 @@
+use super::*;
+
+#[test]
+fn all_zen_files_have_expected_output() {
+    let dir = test_dir();
+    let mut missing = Vec::new();
+    for entry in std::fs::read_dir(&dir).expect("read tests/zen") {
+        let entry = entry.expect("dir entry");
+        let path = entry.path();
+        if path.extension().and_then(|s| s.to_str()) == Some("zen") {
+            let stem = path.file_stem().unwrap().to_str().unwrap();
+            let expected = dir.join("expected").join(format!("{}.expected", stem));
+            if !expected.exists() {
+                missing.push(stem.to_string());
+            }
+        }
+    }
+    assert!(
+        missing.is_empty(),
+        "missing .expected files for: {:?}",
+        missing
+    );
+}
